wifi: add firmware loader support for the non-LKM case
if the wlan module is built into the kernel and the device doesn't
provide a wlan_mac.bin on the persist partition, then that file has
to be generated from NV data.

There currently isn't any way to let the wifi HAL wait for that file though.
This patch reuses the existing firmware loader code to launch a process
which can create the file with the mac address.

Change-Id: I21a4a74e1a48894833e641cbc34fbfb9316c9151
Signed-off-by: Corey Edwards <ensabahnur16@gmail.com>